+++ front/src/chunks/index.js
@@ -1,0 +1,3 @@
+export { default as ChatBox } from './ChatBox';
+export { default as Participants } from './pages/Participants';
+export { default as TagList } from './pages/TagList';
