+++ front/src/components/index.js
@@ -1,0 +1,23 @@
+export { default as Avatar } from './Avatar/Avatar';
+export { default as AvatarImage } from './Avatar/AvatarImage';
+export { default as AvatarWithBadge } from './Avatar/AvatarWithBadge';
+export { default as Badge } from './Badge/Badge';
+export { default as BadgeClickable } from './Badge/BadgeClickable';
+export { default as RoundButton } from './Button/RoundButton';
+export { default as SquareButton } from './Button/SquareButton';
+export { default as ChattingForm } from './ChattingForm/ChattingForm';
+export { default as GameCard } from './GameCard/GameCard';
+export { default as MainImage } from './MainImage/MainImage';
+export { default as Modal } from './Modal/Modal';
+export { default as ModalError } from './Modal/ModalError';
+export { default as ModalMinimized } from './Modal/ModalMinimized';
+export { default as NicknameSection } from './NicknameSection/NicknameSection';
+export { default as Room } from './Room/Room';
+export { default as SearchInput } from './SearchInput/SearchInput';
+export { default as DropdownInput } from './SearchInput/DropdownInput';
+export { default as TextInput } from './SearchInput/TextInput';
+export { default as Slider } from './Slider/Slider';
+export { default as SpeechBubble } from './SpeechBubble/SpeechBubble';
+export { default as SpeechBubbleWithAvatar } from './SpeechBubble/SpeechBubbleWithAvatar';
+export { default as Tag } from './Tag/Tag';
+export { default as TagErasable } from './Tag/TagErasable';
